requirements.txt
aiofiles==23.2.1
aiogram==3.5.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.1
asn1crypto==1.5.1
asyncio==3.4.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
css==0.1
Django==5.0.4
djangorestframework==3.15.2
environs==11.0.0
frozenlist==1.4.1
html5lib==1.1
idna==3.7
iniconfig==2.0.0
magic-filter==1.0.12
marshmallow==3.22.0
mirakuru==2.5.2
multidict==6.0.5
packaging==24.0
pg8000==1.31.1
pluggy==1.5.0
port-for==0.7.2
postgresql-api==2.0.2.dev1
postgresqlite==0.4.3
prettier==0.0.7
psutil==5.9.8
psycopg==3.1.18
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pyOpenSSL==24.1.0
pyTelegramBotAPI==4.17.0
pytest==8.1.1
pytest-postgresql==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
scramp==1.4.5
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
TelegramBot-API==0.0.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
yarl==1.9.4